extracted MobileAccordion Component
