+++ src/app/[locale]/product/_components/ProductOverviewCard.tsx
@@ -1,7 +1,4 @@
 import {
-    Accordion,
-    AccordionDetails,
-    AccordionSummary,
     Box,
     Card,
     CardContent,
@@ -10,7 +7,6 @@ import {
 } from '@mui/material';
 import React, { useEffect, useState } from 'react';
 import { DataRow } from 'components/basics/DataRow';
-import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
 import { AssetAdministrationShell, SubmodelElementCollection } from '@aas-core-works/aas-core3.0-typescript/types';
 import { IconCircleWrapper } from 'components/basics/IconCircleWrapper';
 import { AssetIcon } from 'components/custom-icons/AssetIcon';
@@ -28,6 +24,7 @@ import {
     findValueByIdShort,
 } from 'lib/util/SubmodelResolverUtil';
 import { useProductImageUrl } from 'lib/hooks/UseProductImageUrl';
+import { MobileAccordion } from 'components/basics/detailViewBasics/MobileAccordion';
 
 type ProductOverviewCardProps = {
     readonly aas: AssetAdministrationShell | null;
@@ -37,12 +34,6 @@ type ProductOverviewCardProps = {
     readonly isAccordion: boolean;
     readonly imageLinksToDetail?: boolean;
     readonly repositoryURL: string | null;
-};
-
-type MobileAccordionProps = {
-    readonly content: React.ReactNode;
-    readonly title: string;
-    readonly icon: React.ReactNode;
 };
 
 type ProductClassification = {
@@ -59,20 +50,6 @@ type OverviewData = {
     readonly productClassifications?: ProductClassification[];
 };
 
-function MobileAccordion(props: MobileAccordionProps) {
-    return (
-        <Accordion disableGutters elevation={0} style={{ width: '100%' }}>
-            <AccordionSummary expandIcon={<ArrowDropDownIcon sx={{ color: 'grey.600' }} />}>
-                <Box display="flex" alignItems="center" data-testid="mobile-accordion-header">
-                    <IconCircleWrapper sx={{ mr: 1 }}>{props.icon}</IconCircleWrapper>
-                    <Typography>{props.title}</Typography>
-                </Box>
-            </AccordionSummary>
-            <AccordionDetails data-testid="mobile-accordion-content">{props.content}</AccordionDetails>
-        </Accordion>
-    );
-}
-
 export function ProductOverviewCard(props: ProductOverviewCardProps) {
     const isAccordion = props.isAccordion;
     const navigate = useRouter();
@@ -82,6 +59,7 @@ export function ProductOverviewCard(props: ProductOverviewCardProps) {
     const locale = useLocale();
     const productImageUrl = useProductImageUrl(props.aas, props.repositoryURL);
 
+    // TODO: once we know how to display this data, split it into its own components
     useEffect(() => {
         if (props.submodels && props.submodels.length > 0) {
             const technicalDataSubmodelElements = findSubmodelByIdOrSemanticId(
